Implement CourseController

+++ src/main/java/dev/patika/hw02/business/concretes/CourseManager.java
@@ -1,21 +1,31 @@
 package dev.patika.hw02.business.concretes;
 
 import dev.patika.hw02.business.abstracts.CourseService;
+import dev.patika.hw02.business.abstracts.InstructorService;
+import dev.patika.hw02.core.exceptions.EntityNotExistsException;
+import dev.patika.hw02.core.exceptions.ForeignKeyConstraintViolationException;
+import dev.patika.hw02.core.exceptions.UniqueConstraintViolationException;
 import dev.patika.hw02.dataaccess.abstracts.CourseRepository;
 import dev.patika.hw02.entities.concretes.Course;
+import dev.patika.hw02.entities.concretes.Instructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.util.Pair;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class CourseManager implements CourseService {
 
     private final CourseRepository repository;
+    private final InstructorService instructorService;
 
     @Autowired
-    public CourseManager(CourseRepository repository) {
+    public CourseManager(CourseRepository repository, InstructorService instructorService) {
         this.repository = repository;
+        this.instructorService = instructorService;
     }
 
     @Override
@@ -29,17 +39,57 @@ public class CourseManager implements CourseService {
     }
 
     @Override
+    public Course findByCode(String code) {
+        return repository.findByCode(code);
+    }
+
+    @Override
     public Course create(Course course) {
+
+        // check if 'code' is unique
+        Course existsCourse = findByCode(course.getCode());
+        if (existsCourse != null)
+            throw new UniqueConstraintViolationException("code", course.getCode());
+
+        // Check if there is an instructor with the foreign key 'instructorId'
+        Instructor existsInstructor = instructorService.findById(course.getInstructor().getId());
+        if (existsInstructor == null)
+            throw new ForeignKeyConstraintViolationException("instructorId", course.getInstructor().getId());
+
         return repository.save(course);
     }
 
     @Override
     public Course update(Course course) {
+
+        // Check if the course is exists
+        if (findById(course.getId()) == null)
+            throw new EntityNotExistsException("Course", Pair.of("id", course.getId()));
+
+        // check if 'code' is unique
+        Course existsCourse = findByCode(course.getCode());
+        if (existsCourse != null && !Objects.equals(existsCourse.getId(), course.getId()))
+            throw new UniqueConstraintViolationException("code", course.getCode());
+
+        // Check if there is an instructor with the foreign key 'instructorId'
+        Instructor existsInstructor = instructorService.findById(course.getInstructor().getId());
+        if (existsInstructor == null)
+            throw new ForeignKeyConstraintViolationException("instructorId", course.getInstructor().getId());
+
         return repository.update(course);
     }
 
+    @Transactional
     @Override
     public void deleteById(Long id) {
+
+        Course course = findById(id);
+        if (course == null)
+            throw new EntityNotExistsException("Course", Pair.of("id", id));
+
+        course.clearStudents();
+
+        repository.update(course);
         repository.deleteById(id);
     }
 }
